requirements.txt
bcrypt==3.2.0
boto3
channels
daphne
Django==4.2.4
django-allauth
django-cors-headers
django-filter
django-dotenv==1.4.2
django-storages[google]
google-cloud-storage
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-anymail
# imageio==2.31.1
# imageio-ffmpeg==0.4.8
# jwcrypto==1.5.0
# moviepy==1.0.3
# oauthlib==3.2.0
# passlib==1.7.4
# Pillow==9.0.1
Pillow
psycopg
# psycopg2
# psycopg2-binary==2.9.6
# pydub
PyJWT
# python-decouple==3.8
python-dotenv==1.0.0
# requests>=2.25.1,<5
# requests-oauthlib==1.3.1
# requests-toolbelt==0.9.1
# Routes==2.5.1
redis
# Production server and database utilities
uvicorn[standard]
dj-database-url
